% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic.saturation.R
\name{acoustic.saturation}
\alias{acoustic.saturation}
\title{Acoustic saturation}
\usage{
acoustic.saturation(pow, bgn, min.db = -90, max.db = 6.5, p = 0.9,
  noise = NULL, decibels = FALSE, raw.freq = FALSE, probs = FALSE,
  smooth = FALSE, window = 5)
}
\arguments{
\item{pow}{NxN matrix of Power (POW) values (must match dimensions of bgn)}

\item{bgn}{NxN matrix of Background Noise (BGN) values (must match dimensions of pow)}

\item{min.db}{Minimum decibels (default is 0)}

\item{max.db}{Maximum decibels (default is 6.5)}

\item{p}{Threshold percentile for bgn, default is p=0.90 (90th percentile)}

\item{noise}{Remove noise in the BGN. Default is null and valid options are mean or median)}

\item{decibels}{(FALSE/TRUE) If TRUE, transform the data, assuming Hz, to decibels using Db=20*log(x)}

\item{raw.freq}{(FALSE/TRUE) Return raw frequencies rather than normalized}

\item{probs}{(FALSE/TRUE) Return a vector of row-wise probabilities indicating saturation}

\item{smooth}{(FALSE/TRUE) Smooth the raw or normalized frequencies}

\item{window}{Size of window to use in smoothing}
}
\value{
A vector equal to nrow(pow & bgn)
}
\description{
Calculates the acoustic saturation index (saturation per minute/row).
}
\examples{
 # Small example
 POW <- matrix(runif(2000, 0, 15), nrow = 100, ncol = 20)
 BGN <- matrix(runif(2000, 0, 15), nrow = 100, ncol = 20)

 # Normalized (raw.freq = FALSE) and smoothed (smooth=TRUE)	
 summary( sat <- acoustic.saturation(POW, BGN, smooth = TRUE) )

 # Return probability of saturation	
 ( sat <- acoustic.saturation(POW, BGN, probs = TRUE) )

 # Large example
 rtnorm <- function(n, mean, sd, a = -Inf, b = Inf){
     qnorm(runif(n, pnorm(a, mean, sd), pnorm(b, mean, sd)), mean, sd)
 }
 POW <- matrix(rtnorm(368640, mean = 1.60741, sd = 1.66311, a = 0, b = 23.4575), 
               nrow = 1440, ncol = 256) 
 BGN <- matrix(rtnorm(368640, mean = 1.60741, sd = 1.66311, a = 0, b = 23.4575), 
               nrow = 1440, ncol = 256) 
 
 # Normalized (raw.freq = FALSE) and smoothed (smooth=TRUE)			  
 summary( sat <- acoustic.saturation(POW, BGN, smooth = TRUE) )
 
 # Not normalized (raw.freq = TRUE) and not smoothed (smooth=FALSE)		
 summary( sat <- acoustic.saturation(POW, BGN, raw.freq = TRUE, smooth = FALSE) )

}
\references{
Burivalova, Z., M. Towsey, T. Boucher, A. Truskinger, P. Roe & E.T. Game (2018) Using 
              soundscapes to detect variable degrees of human influence on tropical forests in 
              Papua New Guinea. Conservation Biology 329(1):205-215
}
\seealso{
\code{\link[smooth]{sma}} for details on ARIMA smoothing
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org> and Tim Boucher <tboucher@tnc.org>
}
